pyproject.toml
[tool.poetry]
name = "curse-api"
repository = "https://github.com/Stinky-c/curse-api"
homepage = "https://github.com/Stinky-c/curse-api"
readme = "README.md"
version = "2.0.0"
description = "A simple curseforge api wrapper"
authors = ["Stinky-c <60587749+Stinky-c@users.noreply.github.com>"]
license = "MIT"
keywords = ["curseforge", "api wrapper"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "License :: OSI Approved :: MIT License",
]


[tool.poetry.dependencies]
python = "^3.7"
pydantic = "^1.10.4"

httpx = { version = "^0.23.1", optional = true }
aiohttp = { version = "^3.8.4", optional = true }

[tool.poetry.extras]
httpx = ["httpx"]
aiohttp = ["aiohttp"]
quicks = ["httpx"]


[tool.poetry.group.dev.dependencies]
python-dotenv = "^0.21.0"
pytest = "^7.1.3"
tox = "^3.26.0"
black = { version = "^22.10.0", allow-prereleases = true }
pytest-asyncio = "^0.20.3"
ruff = "^0.0.259"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
extend-exclude = [".venv", "examples", "tests"]
ignore-init-module-imports = true
extend-ignore = ["E501"]
extend-select = ["T201"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["E402", "F401"]
